refactor: move create project handler
