+++ packages/vscode-extension/src/handlers.ts
@@ -135,6 +135,7 @@ import { getSystemInputs } from "./utils/systemEnvUtils";
 import { getTriggerFromProperty, isTriggerFromWalkThrough } from "./utils/telemetryUtils";
 import { openFolder, openOfficeDevFolder } from "./utils/workspaceUtils";
 import { openWelcomeHandler } from "./handlers/openLinkHandlers";
+import { createNewProjectHandler } from "./handlers/lifecycleHandlers";
 
 export function activate(): Result<Void, FxError> {
   const result: Result<Void, FxError> = ok(Void);
@@ -336,42 +337,6 @@ async function refreshEnvTreeOnFileContentChanged(workspacePath: string, filePat
   if (path.normalize(filePath) === path.normalize(projectSettingsPath)) {
     await envTreeProviderInstance.reloadEnvironments();
   }
-}
-
-export async function createNewProjectHandler(...args: any[]): Promise<Result<any, FxError>> {
-  ExtTelemetry.sendTelemetryEvent(TelemetryEvent.CreateProjectStart, getTriggerFromProperty(args));
-  let inputs: Inputs | undefined;
-  if (args?.length === 1) {
-    if (!!args[0].teamsAppFromTdp) {
-      inputs = getSystemInputs();
-      inputs.teamsAppFromTdp = args[0].teamsAppFromTdp;
-    }
-  } else if (args?.length === 2) {
-    // from copilot chat
-    inputs = { ...getSystemInputs(), ...args[1] };
-  }
-  const result = await runCommand(Stage.create, inputs);
-  if (result.isErr()) {
-    return err(result.error);
-  }
-
-  const res = result.value as CreateProjectResult;
-  if (res.shouldInvokeTeamsAgent) {
-    await invokeTeamsAgent([TelemetryTriggerFrom.CreateAppQuestionFlow]);
-    return result;
-  }
-  const projectPathUri = Uri.file(res.projectPath);
-  // If it is triggered in @office /create for code gen, then do no open the temp folder.
-  if (isValidOfficeAddInProject(projectPathUri.fsPath) && inputs?.agent === "office") {
-    return result;
-  }
-  // show local debug button by default
-  if (isValidOfficeAddInProject(projectPathUri.fsPath)) {
-    await openOfficeDevFolder(projectPathUri, true, res.warnings, args);
-  } else {
-    await openFolder(projectPathUri, true, res.warnings, args);
-  }
-  return result;
 }
 
 export async function selectAndDebugHandler(args?: any[]): Promise<Result<null, FxError>> {
